backend/requirements.txt
asgiref==3.5.2
click==8.1.3
colorama==0.4.6
Django==4.1.4
djangorestframework==3.14.0
joblib==1.2.0
nltk==3.7
pyphen==0.13.2
pytz==2022.6
regex==2022.10.31
sqlparse==0.2.4
tqdm==4.64.1
tzdata==2022.7
tokenizers==0.13.2
transformers==4.25.1
torch==1.13.0
wordfreq==3.0.3
pymongo == 4.3.3
tk==0.1.0
PyPDF2 == 3.0.1
docx2txt == 0.8
ebooklib==0.18
bs4==0.0.1
drf-yasg==1.21.4
gunicorn==20.1.0
djongo==1.3.6
pytest-django==4.5.2